refactor(client): extract every global e2e method inside a mpe-e2e-utils-ts file
